Moved add/remove market logic to MarketParticipant trait.
